+++ crates/tauri-plugin-holochain-service-consumer/android/src/main/java/HolochainServiceConsumerPlugin.kt
@@ -2,6 +2,11 @@ package com.plugin.holochain_service_consumer
 
 import android.app.Activity
 import android.webkit.WebView
+import android.util.Log
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.SupervisorJob
 import app.tauri.annotation.Command
 import app.tauri.annotation.TauriPlugin
 import app.tauri.plugin.JSObject
@@ -9,23 +14,30 @@ import app.tauri.plugin.Plugin
 import app.tauri.plugin.Invoke
 import org.holochain.androidserviceruntime.holochain_service_client.HolochainServiceClient
 import org.holochain.androidserviceruntime.holochain_service_client.toJSONObjectString
-import org.holochain.androidserviceruntime.holochain_service_client.toJSONArrayString
 
 @TauriPlugin
-class HolochainServicePlugin(private val activity: Activity): Plugin(activity) {
+class HolochainServiceConsumerPlugin(private val activity: Activity): Plugin(activity) {
     private lateinit var webView: WebView
     private lateinit var injectHolochainClientEnvJavascript: String
-    private lateinit var serviceClient: HolochainServiceClient
+    private val supervisorJob = SupervisorJob()
+    private val serviceScope = CoroutineScope(supervisorJob)
+    private var serviceClient = HolochainServiceClient(
+        this.activity,
+        "org.holochain.androidserviceruntime.app",
+        "com.plugin.holochain_service.HolochainService"
+    )
+    private var TAG = "HolochainServiceConsumerPlugin"
 
     /**
      * Load the plugin, start the service
      */
     override fun load(webView: WebView) {
+        Log.d(TAG, "load")
         super.load(webView)
         this.webView = webView
 
         // Load holochain client injected javascript from resource file
-        val resourceInputStream = this.activity.resources.openRawResource(R.raw.injectholochainclientenv)
+        val resourceInputStream = this.activity.resources.openRawResource(R.raw.holochainenv)
         this.injectHolochainClientEnvJavascript = resourceInputStream.bufferedReader().use { it.readText() }
     }
 
@@ -34,11 +46,7 @@ class HolochainServicePlugin(private val activity: Activity): Plugin(activity) {
      */
     @Command
     fun connect(invoke: Invoke) {
-        this.serviceClient = HolochainServiceClient(
-            this.activity,
-            "com.plugin.holochain_service.HolochainService",
-            "org.holochain.androidserviceruntime.app"
-        )
+        Log.d(TAG, "connect")
         this.serviceClient.connect()
         invoke.resolve()
     }
@@ -48,9 +56,25 @@ class HolochainServicePlugin(private val activity: Activity): Plugin(activity) {
      */
     @Command
     fun installApp(invoke: Invoke) {
+        Log.d(TAG, "installApp")
         val args = invoke.parseArgs(InstallAppPayloadFfiInvokeArg::class.java)
-        val res = this.serviceClient.installApp(args.toFfi())
-        invoke.resolve(JSObject(res.toJSONObjectString()))
+        serviceScope.launch(Dispatchers.IO) {
+            val res = serviceClient.installApp(args.toFfi())
+            invoke.resolve(JSObject(res.toJSONObjectString()))
+        }
+    }
+
+    /**
+     * Enable an app
+     */
+    @Command
+    fun enableApp(invoke: Invoke) {
+        Log.d(TAG, "enableApp")
+        val args = invoke.parseArgs(AppIdInvokeArg::class.java)
+        serviceScope.launch(Dispatchers.IO) {
+            val res = serviceClient.enableApp(args.installedAppId)
+            invoke.resolve(JSObject(res.toJSONObjectString()))
+        }
     }
 
     /**
@@ -58,11 +82,14 @@ class HolochainServicePlugin(private val activity: Activity): Plugin(activity) {
      */
     @Command
     fun isAppInstalled(invoke: Invoke) {
+        Log.d(TAG, "isAppInstalled")
         val args = invoke.parseArgs(AppIdInvokeArg::class.java)
-        val res = this.serviceClient.isAppInstalled(args.installedAppId)
-        val obj = JSObject()
-        obj.put("installed", res)
-        invoke.resolve(obj)
+        serviceScope.launch(Dispatchers.IO) {
+            val res = serviceClient.isAppInstalled(args.installedAppId)
+            val obj = JSObject()
+            obj.put("installed", res)
+            invoke.resolve(obj)
+        }
     }
 
     /**
@@ -71,9 +98,12 @@ class HolochainServicePlugin(private val activity: Activity): Plugin(activity) {
     @OptIn(ExperimentalUnsignedTypes::class)
     @Command
     fun ensureAppWebsocket(invoke: Invoke) {
+        Log.d(TAG, "ensureAppWebsocket")
         val args = invoke.parseArgs(AppIdInvokeArg::class.java)
-        val res = this.serviceClient.ensureAppWebsocket(args.installedAppId)
-        invoke.resolve(JSObject(res.toJSONObjectString()))
+        serviceScope.launch(Dispatchers.IO) {
+            val res = serviceClient.ensureAppWebsocket(args.installedAppId)
+            invoke.resolve(JSObject(res.toJSONObjectString()))
+        }
     }
 
     /**
@@ -81,8 +111,11 @@ class HolochainServicePlugin(private val activity: Activity): Plugin(activity) {
      */
     @Command
     fun signZomeCall(invoke: Invoke) {
+        Log.d(TAG, "signZomeCall")
         val args = invoke.parseArgs(ZomeCallUnsignedFfiInvokeArg::class.java)
-        val res = this.serviceClient.signZomeCall(args.toFfi())
-        invoke.resolve(JSObject(res.toJSONObjectString()))
+        serviceScope.launch(Dispatchers.IO) {
+            val res = serviceClient.signZomeCall(args.toFfi())
+            invoke.resolve(JSObject(res.toJSONObjectString()))
+        }
     }
 }