+++ crates/tauri-plugin-holochain-service-consumer/guest-js/tauri-commands/index.ts
@@ -16,6 +16,6 @@ export async function isAppInstalled(installedAppId: string): Promise<boolean> {
   return await invoke<{installed: boolean}>('plugin:holochain-service-consumer|is_app_installed', { installedAppId }).then((r) => (r.installed));
 }
 
-export async function ensureAppWebsocket(installedAppId: string): Promise<{installedAppId: string, port: number, token: Uint8Array} | null> {
-  return await invoke<{installedAppId: string, port: number, token: Uint8Array}>('plugin:holochain-service-consumer|ensure_app_websocket', { installedAppId });
+export async function ensureAppWebsocket(installedAppId: string): Promise<{port: number, authentication: {token: Uint8Array, expiresAt?: number}} | null> {
+  return await invoke<{port: number, authentication: {token: Uint8Array, expiresAt?: number}}>('plugin:holochain-service-consumer|ensure_app_websocket', { installedAppId });
 }