+++ crates/tauri-plugin-holochain-service/android/src/main/java/com/plugin/holochain_service/HolochainServicePlugin.kt
@@ -3,7 +3,9 @@ package com.plugin.holochain_service
 import android.app.Activity
 import android.app.NotificationChannel
 import android.app.NotificationManager
+import android.content.ComponentName
 import android.content.Context
+import android.content.Intent
 import android.webkit.WebView
 import android.util.Log
 import kotlinx.coroutines.launch
@@ -24,10 +26,16 @@ import org.holochain.androidserviceruntime.holochain_service_client.toJSONArrayS
 class HolochainServicePlugin(private val activity: Activity): Plugin(activity) {
     private lateinit var webView: WebView
     private lateinit var injectHolochainClientEnvJavascript: String
-    private lateinit var serviceClient: HolochainServiceClient
+    private val packageName = "org.holochain.androidserviceruntime.app"
+    private val className = "com.plugin.holochain_service.HolochainService"
+    private var serviceClient = HolochainServiceClient(
+        this.activity,
+        this.packageName,
+        this.className,
+    )
     private val supervisorJob = SupervisorJob()
     private val serviceScope = CoroutineScope(supervisorJob)
-    private var TAG = "HolochainServicePlugin"
+    private val TAG = "HolochainServicePlugin"
 
     /**
      * Load the plugin, start the service
@@ -38,7 +46,7 @@ class HolochainServicePlugin(private val activity: Activity): Plugin(activity) {
         this.webView = webView
 
         // Load holochain client injected javascript from resource file
-        val resourceInputStream = this.activity.resources.openRawResource(R.raw.injectholochainclientenv)
+        val resourceInputStream = this.activity.resources.openRawResource(R.raw.holochainenv)
         this.injectHolochainClientEnvJavascript = resourceInputStream.bufferedReader().use { it.readText() }
 
         // Create notification channel
@@ -56,11 +64,12 @@ class HolochainServicePlugin(private val activity: Activity): Plugin(activity) {
     @Command
     fun start(invoke: Invoke) {
         Log.d(TAG, "start")
-        this.serviceClient = HolochainServiceClient(
-            this.activity,
-            "com.plugin.holochain_service.HolochainService",
-            "org.holochain.androidserviceruntime.app"
-        )
+
+        // Start service
+        val intent = Intent()
+        intent.setComponent(ComponentName(this.packageName, this.className))
+        this.activity.startForegroundService(intent)
+
         this.serviceClient.connect()
         invoke.resolve()
     }