+++ examples/service-app/src/Labelled.svelte
@@ -1,8 +1,0 @@
-<script lang="ts">
-  export let label;
-</script>
-
-<div style="text-align: left; margin-bottom: 10px">
-  <b>{label}</b>
-  <div><slot></slot></div>
-</div>