+++ CHANGELOG.md
@@ -18,3 +18,5 @@
 - Support Json serialization of sealed classes, cleanup and tests of Json serialization.
 - Fix inconsistent crashes on relaunch with "logger already initialized" errors.
 - Add nix flake for android + holochain development, remove reliance on p2p-shipyard flake.
+- Added example app demonstrating use of tauri-plugin-holochain-service-consumer
+